backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
Markdown==3.7
pillow==11.1.0
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.1
djangorestframework-recaptcha